requirements.txt
setuptools~=49.2.1
requests~=2.25.1
torch~=1.7.1
nltk~=3.5
sklearn
scikit-learn~=0.23.2
itu.algs4